requirements.txt
django
djangorestframework
django-cors-headers
dj-database-url 
python-decouple 
whitenoise
